requirements.txt
fastapi>=0.115.2
uvicorn>=0.27.0
pydantic>=2.5.3
python-dotenv==1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
openai>=1.12.0
chromadb==0.4.22
tiktoken>=0.9.0
python-multipart>=0.0.18
pypdf2==3.0.1
unstructured==0.10.30
pdf2image==1.16.3
pytesseract==0.3.10
pydantic-settings>=2.0.0
starlette>=0.46.1